Add Links Between SCRIPTs & USE/STOP
In §3.7 *SCRIPTs* add links to USE and STOP statements when mentioned in
text; although the subjects are strictly realated they are far apart from
each other in the Manual, so cross-reference links are most useful here
for the reader.

And, viceversa, in §3.17 *Actor Statements* > *USE Statement*, add a
link to *SCRIPTs*, to allow reader to easily navigate back & forth
between these two closely-related sections.