Switch to Dart Sass
As the end-of-life of Ruby Sass is approaching (26 March 2019), the
project now switches to using the new Dart Sass implementation.

Some scripts needed to be tweaked (the "scss" command is no longer an
alias of "sass"), and the compiled stylesheets are slightly different.
Also, the output CSS file now have Unix style LF EOL.