B.Guide 0.1.12
Plasma PNG: Replace `plasma.jpg` with `plasma.png` that has transparency
around the Tv image (knocked background) so that it blends well both in
the document transcript (which has blue BG color) as well as Alan terps.

- Fix `plasma.png` in tutorial game-transcript (§19. Turn It On).
- Replace all occurences of `plasma.jpg` with `plasma.png` in document
  code and in `tvtime.alan.`

Retouch formatting of `tvtime.alan`, regenerate UTF-8 version.
Add linenumbers to final adventure in document (for editing purposes,
might be removed in final document, or not).