B.Guide: Polish Libray Sources
This commit polished the Lib files in `alanguide-code/lib/` by enforcing
prose-casing, a consistent style, better indentation and a few small
fixes to update obsolete reference links and info.

All the changes are now mentioned in the README file inside the Lib
folder (now in markdown format).

The same polishes and changes are then replayed over the tutorial copy
of the lib folder, `alanguide-code/mylib/` while preserving the changes
applied by the tutorial steps. Prose casing is then applied to the
tweaked code and added files in `mylib/`.

All Alan-UTF8 files are regenerated, and the B.Guide is rebuilt to
incorporate the changes in the imported sources.

B.Guide is bumped up to v0.1.13.