Merge branch 'handle_elisions'
In view of the upcoming Alan 3.0 beta7 release, we shall do all changes
to the Manual in this `beta7-prep` development branch, which we'll then
squash into master once beta7 is publicly available.

Merge branch `handle_elisions` into `beta7-prep` and resolve conflicts
on Chapter 5 source by using "theirs" stategy.

(See #40 for a discussion on this)