Fix Section Level
Moved one level up sub-section of §3.17, *Repetition Statements*, which
was wrongly portraided as a subsection of *Actor Statements*.

Also, change Manual version to v0.1.91-PreReleaseBeta7, which will be
kept as temporary development version until ready to merge into `master`
when Alan Beta7 is ready (it will become v0.1.91).

From now only the date of the manual should be changed.