Fix Section Numbering in PDF
In order to preserve same section numbering across backends, remove
the section numbering suppression in "§3.1. General Rules" so that
both HTML and PDF docs have same numbering. (Fixes #68)

Unfortunately, the section numbering remains in the two PDF appendices:

- "Appendix E: Predefined Player Words"
- "Appendix I: Copying Conditions"

But these don't have a negative impact on the different editions, they
are just spurious numberings in the PDF which are not preserved in the
HTML edition. So, no harm done.