Haml Listing Template v1.1.0
Fix `block_listing.html.haml`, the Haml HTML listing template used for
integrating Highlight into the Asciidoctor toolchain.

The original template was chewing up indentation of the first line in
code blocks that began with an indented line. This commit fixes the
problem by using the `:preserve` filter to ensure that indentation is
never affected. (Fixes #70)