[manual] Adjust Diagram With
Explicitly set to 90% the scale with of the SVG diagram of §2.4, Fig.2
(Predefined Classes) to ensure correct handling of the image after the
latest changes to the Dia export filter (now using Cairo-SVG).

The new scale/width settings should cover most backends and output
formats (tested with HTML5 and PDF via asciidoctor-fopub).