New Status Doc
Create `STATUS.md` doc summarizing the current status of the various
syntax highlighters available in this project, and their limits in terms
of supported and unsupported features, providing a lookup table for
quick consultation.

In the future this doc will also provide status info about all the docs
in this project, and which highlighter they are using (see #118).